requirements.txt
python_dotenv==0.6.3
requests==2.9.1
tqdm==4.11.2
